LayoutTests/imported/w3c/web-platform-tests/editing/other/shadow_root_iteration_in_block_style-expected.txt

helloworld

FAIL RemoveFormat should remove CSS styles of selected node assert_equals: expected "hello<span id=\"span\" contenteditable=\"\" style=\"\">world</span>" but got "hello<span id=\"span\" contenteditable=\"\">world</span>"
